pyproject.toml
[project]
name = "image-preprocessing-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "numpy>=2.4.1",
    "openai>=2.15.0",
    "pillow>=12.1.0",
    "python-dotenv>=1.2.1",
    "swarm",
]

[tool.uv.sources]
swarm = { git = "https://github.com/openai/swarm.git" }
